fix: move JobMonitorService to execution package, fix circular dependency
- Move job_monitor_service.py from services/ to execution/ (queries
  ActorRegistry which lives in execution, so it belongs there)
- Move display_active_jobs from ServeStatusService to server/app.py
  (CLI layer can import from both execution and services)
- Update all imports and exports accordingly
- Fix circular dependency: execution -> services -> execution
