+++ src/vibe3/services/__init__.py
@@ -60,11 +60,6 @@ if TYPE_CHECKING:
     from vibe3.services.issue_branch_resolver import resolve_issue_branch_input
 
     # Job monitoring
-    from vibe3.services.job_monitor_service import (
-        ActiveJob,
-        JobMonitorService,
-        JobMonitorSnapshot,
-    )
     from vibe3.services.label_service import LabelService
     from vibe3.services.orchestra_helpers import (
         get_handoff_state_label,
@@ -161,7 +156,6 @@ if TYPE_CHECKING:
     from vibe3.services.verdict_service import VerdictService
 
 __all__ = [
-    "ActiveJob",
     "BaseResolutionUsecase",
     "BlockedStateService",
     "BootstrapAction",
@@ -192,8 +186,6 @@ __all__ = [
     "IssueFlowService",
     "IssueStatusEntry",
     "IssueTitleCacheService",
-    "JobMonitorService",
-    "JobMonitorSnapshot",
     "LabelAnomaly",
     "LabelService",
     "MissingTaskIssueError",
@@ -295,9 +287,6 @@ _SYMBOL_MODULES = {
     "ExpiredResourceCleanupService": "vibe3.services.expired_resource_cleanup_service",
     "FlowCategory": "vibe3.services.flow_classifier",
     # Job monitoring
-    "ActiveJob": "vibe3.services.job_monitor_service",
-    "JobMonitorService": "vibe3.services.job_monitor_service",
-    "JobMonitorSnapshot": "vibe3.services.job_monitor_service",
     "FlowCleanupService": "vibe3.services.flow_cleanup_service",
     "FlowOrchestratorService": "vibe3.services.flow_orchestrator_service",
     "FlowProjection": "vibe3.services.flow_projection_service",
