+++ src/vibe3/services/serve_status_service.py
@@ -49,7 +49,6 @@ class ServeStatusService:
         self._display_recent_activity()
         self._display_failed_gate()
         self._display_error_tracking()
-        self._display_active_jobs()
 
     @staticmethod
     def _clean_error_message(error_message: str, max_length: int = 100) -> str:
@@ -257,72 +256,3 @@ class ServeStatusService:
                 self.console.print(table)
         else:
             self.console.print("  No errors recorded")
-
-    def _display_active_jobs(self) -> None:
-        """Display active and recent jobs from the actor registry."""
-        from vibe3.services.job_monitor_service import JobMonitorService
-
-        service = JobMonitorService()
-        snapshot = service.snapshot()
-
-        self.console.print("[bold]Active Jobs:[/bold]")
-
-        if not snapshot.active_jobs and not snapshot.recent_jobs:
-            self.console.print("  No active jobs")
-            self.console.print()
-            return
-
-        if snapshot.active_jobs:
-            table = Table(title="Running / Queued", show_lines=True)
-            table.add_column("Actor", style="cyan", width=20)
-            table.add_column("Type", style="magenta", width=12)
-            table.add_column("Status", style="green", width=8)
-            table.add_column("Issue", style="yellow", width=10)
-            table.add_column("Branch", style="white")
-            table.add_column("Started", style="dim", width=12)
-
-            for job in snapshot.active_jobs:
-                started_display = (
-                    format_age_aware_time(job.started_at) if job.started_at else "-"
-                )
-                issue_display = f"#{job.issue_number}" if job.issue_number > 0 else "-"
-                table.add_row(
-                    job.actor_id[:20],
-                    job.job_type.upper(),
-                    job.status.upper(),
-                    issue_display,
-                    job.branch,
-                    started_display,
-                )
-            self.console.print(table)
-
-        if snapshot.recent_jobs:
-            table = Table(title="Recent (last 30m)", show_lines=True)
-            table.add_column("Actor", style="cyan", width=20)
-            table.add_column("Type", style="magenta", width=12)
-            table.add_column("Status", style="yellow", width=8)
-            table.add_column("Issue", style="yellow", width=10)
-            table.add_column("Branch", style="white")
-            table.add_column("Completed", style="dim", width=12)
-
-            for job in snapshot.recent_jobs:
-                completed_display = (
-                    format_age_aware_time(job.completed_at) if job.completed_at else "-"
-                )
-                issue_display = f"#{job.issue_number}" if job.issue_number > 0 else "-"
-                table.add_row(
-                    job.actor_id[:20],
-                    job.job_type.upper(),
-                    job.status.upper(),
-                    issue_display,
-                    job.branch,
-                    completed_display,
-                )
-            self.console.print(table)
-
-        self.console.print(
-            f"  Summary: {snapshot.running_count} running, "
-            f"{snapshot.completed_count} completed, "
-            f"{snapshot.failed_count} failed (last 30m)"
-        )
-        self.console.print()
